requirements.txt
torch
albumentations
einops>=0.6.1
numpy>=1.24.4
pillow>=9.5.0
omegaconf
pyyaml
pytorch-lightning
monai
tensorboardX
pandas
scikit-learn